Split helpers.ts into merge.ts — under 400 lines each, file-local internal types
Address CodeRabbit P2 findings:
- helpers.ts was 542 lines; split the merge helpers (mergePair,
  mergePairWithQuestion, assignMergeAnswers, submitMerge, getMergeVersion,
  mergeWithAnswerConflict, mergeNonConflictingAnswer) into a new merge.ts
  (240 lines). helpers.ts is now 313 lines — both under the ~400 target.
- Made DirectBooking type and expectMergeAnswer function file-local (removed
  export) — they are only used inside their own module, so per AGENTS.md's
  dead-code rule they should not be part of the public helper API.
- Updated merge-{panel,post,conflicts}.test.ts imports accordingly.
